# frontend/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "costrict",
    "description": "costrict api",
    "version": "0.0.3"
  },
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "config.get",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Get configuration",
        "description": "Retrieve the current CoStrict configuration settings and preferences.",
        "responses": {
          "200": {
            "description": "Get config info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "config.update",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Update configuration",
        "description": "Update CoStrict configuration settings and preferences.",
        "responses": {
          "200": {
            "description": "Successfully updated config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          }
        }
      }
    },
...
    "/session/{sessionID}/message/{messageID}": {
      "get": {
        "operationId": "session.message",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session ID"
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Message ID"
          }
        ],
        "summary": "Get message",
        "description": "Retrieve a specific message from a session by its message ID.",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "parts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Part"
                      }
                    }
                  },
                  "required": [
                    "info",
                    "parts"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Message": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UserMessage"
          },
          {
            "$ref": "#/components/schemas/AssistantMessage"
          }
        ]
      },
...
      "Part": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TextPart"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "sessionID": {
                "type": "string"
              },
              "messageID": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "const": "subtask"
              },
              "prompt": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "agent": {
                "type": "string"
              },
              "command": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "sessionID",
              "messageID",
              "type",
              "prompt",
              "description",
              "agent"
            ]
          },
          {
            "$ref": "#/components/schemas/ReasoningPart"
          },
          {
            "$ref": "#/components/schemas/FilePart"
          },
          {
            "$ref": "#/components/schemas/ToolPart"
          },
          {
            "$ref": "#/components/schemas/StepStartPart"
          },
          {
            "$ref": "#/components/schemas/StepFinishPart"
          },
          {
            "$ref": "#/components/schemas/SnapshotPart"
          },
          {
            "$ref": "#/components/schemas/PatchPart"
          },
          {
            "$ref": "#/components/schemas/AgentPart"
          },
          {
            "$ref": "#/components/schemas/RetryPart"
          },
          {
            "$ref": "#/components/schemas/CompactionPart"
          }
        ]
      },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": {
                "type": "string"
              },
              "additionalProperties": {}
            }
          },
          "success": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "data",
          "errors",
          "success"
        ]
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "const": "NotFoundError"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "name",
          "data"
        ]
      },
...
      "LogLevel": {
        "description": "Log level",
        "type": "string",
        "enum": [
          "DEBUG",
          "INFO",
          "WARN",
          "ERROR"
        ]
      },
      "ServerConfig": {
        "description": "Server configuration for opencode serve and web commands",
        "type": "object",
        "properties": {
          "port": {
            "description": "Port to listen on",
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "hostname": {
            "description": "Hostname to listen on",
            "type": "string"
          },
          "mdns": {
            "description": "Enable mDNS service discovery",
            "type": "boolean"
          },
          "cors": {
            "description": "Additional domains to allow for CORS",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "McpLocalConfig": {
        "type": "object",
        "properties": {
          "type": {
            "description": "Type of MCP server connection",
            "type": "string",
            "const": "local"
          },
          "command": {
            "description": "Command and arguments to run the MCP server",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "environment": {
            "description": "Environment variables to set when running the MCP server",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "string"
            }
          },
          "enabled": {
            "description": "Enable or disable the MCP server on startup",
            "type": "boolean"
          },
          "timeout": {
            "description": "Timeout in ms for fetching tools from the MCP server. Defaults to 5000 (5 seconds) if not specified.",
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "type",
          "command"
        ],
        "additionalProperties": false
      },
...
      "LayoutConfig": {
        "description": "@deprecated Always uses stretch layout.",
        "type": "string",
        "enum": [
          "auto",
          "stretch"
        ]
      },
      "Config": {
        "type": "object",
        "properties": {
          "$schema": {
            "description": "JSON schema reference for configuration validation",
            "type": "string"
          },
          "theme": {
            "description": "Theme name to use for the interface",
            "type": "string"
          },
          "keybinds": {
            "$ref": "#/components/schemas/KeybindsConfig"
          },
          "logLevel": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "tui": {
            "description": "TUI specific settings",
            "type": "object",
            "properties": {
              "scroll_speed": {
                "description": "TUI scroll speed",
                "type": "number",
                "minimum": 0.001
              },
              "scroll_acceleration": {
                "description": "Scroll acceleration settings",
                "type": "object",
                "properties": {
                  "enabled": {
                    "description": "Enable scroll acceleration",
                    "type": "boolean"
                  }
                },
                "required": [
                  "enabled"
                ]
              },
              "diff_style": {
                "description": "Control diff rendering style: 'auto' adapts to terminal width, 'stacked' always shows single column",
                "type": "string",
                "enum": [
                  "auto",
                  "stacked"
                ]
              }
            }
          },
          "server": {
            "$ref": "#/components/schemas/ServerConfig"
          },
          "command": {
            "description": "Command configuration, see https://opencode.ai/docs/commands",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "object",
              "properties": {
                "template": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "agent": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "subtask": {
                  "type": "boolean"
                }
              },
              "required": [
                "template"
              ]
            }
          },
          "watcher": {
            "type": "object",
            "properties": {
              "ignore": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "plugin": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "snapshot": {
            "type": "boolean"
          },
          "share": {
            "description": "Control sharing behavior:'manual' allows manual sharing via commands, 'auto' enables automatic sharing, 'disabled' disables all sharing",
            "type": "string",
            "enum": [
              "manual",
              "auto",
              "disabled"
            ]
          },
          "autoshare": {
            "description": "@deprecated Use 'share' field instead. Share newly created sessions automatically",
            "type": "boolean"
          },
          "autoupdate": {
            "description": "Automatically update to the latest version. Set to true to auto-update, false to disable, or 'notify' to show update notifications",
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string",
                "const": "notify"
              }
            ]
          },
          "disabled_providers": {
            "description": "Disable providers that are loaded automatically",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabled_providers": {
            "description": "When set, ONLY these providers will be enabled. All other providers will be ignored",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "model": {
            "description": "Model to use in the format of provider/model, eg anthropic/claude-2",
            "type": "string"
          },
          "small_model": {
            "description": "Small model to use for tasks like title generation in the format of provider/model",
            "type": "string"
          },
          "default_agent": {
            "description": "Default agent to use when none is specified. Must be a primary agent. Falls back to 'build' if not set or if the specified agent is invalid.",
            "type": "string"
          },
          "username": {
            "description": "Custom username to display in conversations instead of system username",
            "type": "string"
          },
          "mode": {
            "description": "@deprecated Use `agent` field instead.",
            "type": "object",
            "properties": {
              "build": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "plan": {
                "$ref": "#/components/schemas/AgentConfig"
              }
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentConfig"
            }
          },
          "agent": {
            "description": "Agent configuration, see https://opencode.ai/docs/agent",
            "type": "object",
            "properties": {
              "plan": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "build": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "general": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "explore": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "title": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "summary": {
                "$ref": "#/components/schemas/AgentConfig"
              },
              "compaction": {
                "$ref": "#/components/schemas/AgentConfig"
              }
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentConfig"
            }
          },
          "provider": {
            "description": "Custom provider configurations and model overrides",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/ProviderConfig"
            }
          },
          "mcp": {
            "description": "MCP (Model Context Protocol) server configurations",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "anyOf": [
                {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/McpLocalConfig"
                    },
                    {
                      "$ref": "#/components/schemas/McpRemoteConfig"
                    }
                  ]
                },
                {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "enabled"
                  ],
                  "additionalProperties": false
                }
              ]
            }
          },
          "formatter": {
            "anyOf": [
              {
                "type": "boolean",
                "const": false
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "disabled": {
                      "type": "boolean"
                    },
                    "command": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "environment": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "extensions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            ]
          },
          "lsp": {
            "anyOf": [
              {
                "type": "boolean",
                "const": false
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "disabled": {
                          "type": "boolean",
                          "const": true
                        }
                      },
                      "required": [
                        "disabled"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "extensions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "disabled": {
                          "type": "boolean"
                        },
                        "env": {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        "initialization": {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "command"
                      ]
                    }
                  ]
                }
              }
            ]
          },
          "instructions": {
            "description": "Additional instruction files or patterns to include",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "layout": {
            "$ref": "#/components/schemas/LayoutConfig"
          },
          "permission": {
            "$ref": "#/components/schemas/PermissionConfig"
          },
          "tools": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "boolean"
            }
          },
          "enterprise": {
            "type": "object",
            "properties": {
              "url": {
                "description": "Enterprise URL",
                "type": "string"
              }
            }
          },
          "compaction": {
            "type": "object",
            "properties": {
              "auto": {
                "description": "Enable automatic compaction when context is full (default: true)",
                "type": "boolean"
              },
              "prune": {
                "description": "Enable pruning of old tool outputs (default: true)",
                "type": "boolean"
              }
            }
          },
          "experimental": {
            "type": "object",
            "properties": {
              "hook": {
                "type": "object",
                "properties": {
                  "file_edited": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "command": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "environment": {
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "string"
                            }
                          }
                        },
                        "required": [
                          "command"
                        ]
                      }
                    }
                  },
                  "session_completed": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "command": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "environment": {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "command"
                      ]
                    }
                  }
                }
              },
              "chatMaxRetries": {
                "description": "Number of retries for chat completions on failure",
                "type": "number"
              },
              "disable_paste_summary": {
                "type": "boolean"
              },
              "batch_tool": {
                "description": "Enable the batch tool",
                "type": "boolean"
              },
              "openTelemetry": {
                "description": "Enable OpenTelemetry spans for AI SDK calls (using the 'experimental_telemetry' flag)",
                "type": "boolean"
              },
              "primary_tools": {
                "description": "Tools that should only be available to primary agents.",
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "continue_loop_on_deny": {
                "description": "Continue the agent loop when a tool call is denied",
                "type": "boolean"
              },
              "mcp_timeout": {
                "description": "Timeout in milliseconds for model context protocol (MCP) requests",
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              }
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}